organise backend into packages, update references
