+++ base/uk.ac.stfc.isis.ibex.scriptgenerator/src/uk/ac/stfc/isis/ibex/scriptgenerator/PythonBuilder.java
@@ -22,6 +22,7 @@ package uk.ac.stfc.isis.ibex.scriptgenerator;
 import java.util.Collection;
 
 import uk.ac.stfc.isis.ibex.model.ModelObject;
+import uk.ac.stfc.isis.ibex.scriptgenerator.row.Row;
 
 /**
  * Generates Python code based on values in the Script Generator table and the settings.
